Added the ManagerController Feature
