+++ src/main/java/org/highmed/numportal/properties/FeatureProperties.java
@@ -3,7 +3,11 @@ package org.highmed.numportal.properties;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+
 @Data
 @ConfigurationProperties(prefix = "feature")
 public class FeatureProperties {
+
+  private boolean searchWithAql = false;
+
 }
