+++ src/main/resources/application-local.yml
@@ -31,7 +31,7 @@ spring:
           userStoreClient:
             authorization-grant-type: client_credentials
             client-id: num-portal
-            client-secret: ZO29JPc5OLIZ29Rv83tEHOdxiAt2P9eC
+            client-secret: Tp3MiedEuSaLEIxaOhYHsA0bOHQZ7G2C
         provider:
           userStoreClient:
             token-uri: ${keycloak.url}/realms/crr/protocol/openid-connect/token
@@ -107,3 +107,4 @@ user-service:
   delete-users-cron: 0 0 5 * * *
 
 feature:
+  search-with-aql: false
