+++ src/main/resources/application.yml
@@ -206,3 +206,6 @@ pseudonymity:
 
 user-service:
   delete-users-cron: 0 0 5 * * *
+
+feature:
+  search-with-aql: false
